+++ tests/07Templates.spec.js
@@ -20,6 +20,7 @@ test.describe('Templates', () => {
         prepareForSignatureModal,
         templatesPage,
         createNewTemplatePage,
+        successModal,
     }) => {
         await description(
             'Objective: To verify that the user can create a new template in the system successfully. This includes ensuring that all required fields are completed correctly, the template is saved, and it is accessible for future use.'
@@ -40,7 +41,7 @@ test.describe('Templates', () => {
         await prepareForSignatureModal.clickSignOnFieldsMenu();
         await prepareForSignatureModal.clickDocumentBody();
         await prepareForSignatureModal.clickCreateBtn();
-        await prepareForSignatureModal.clickBackToTemplatesBtn();
+        await successModal.clickBackToTemplatesBtn();
 
         await step('Verify that the document status in the template page is "live"', async () => {
             await expect(await templatesPage.table.documentStatus).toHaveText(TEMPLATES_STATUS.live);
@@ -54,6 +55,7 @@ test.describe('Templates', () => {
         templatesPage,
         apiTemplatesPage,
         createNewTemplatePage,
+        successModal,
     }) => {
         await description('Objective: To verify the process of add template to API.');
         await severity(Severity.CRITICAL);
@@ -64,7 +66,7 @@ test.describe('Templates', () => {
 
         test.setTimeout(250 * 1000);
 
-        await createTemplate(signPage, prepareForSignatureModal, templatesPage, createNewTemplatePage);
+        await createTemplate(signPage, prepareForSignatureModal, templatesPage, createNewTemplatePage, successModal);
         await signPage.sideMenu.clickTemplates();
         await templatesPage.table.clickFirstOptionsBtn();
         await templatesPage.table.clickAddToAPIBtn();
@@ -87,6 +89,7 @@ test.describe('Templates', () => {
         editTemplatesPage,
         prepareForSignatureModal,
         createNewTemplatePage,
+        successModal,
     }) => {
         test.slow();
 
@@ -99,7 +102,7 @@ test.describe('Templates', () => {
         await epic('Templates');
         await tags('Edit-template');
 
-        await createTemplate(signPage, prepareForSignatureModal, templatesPage, createNewTemplatePage);
+        await createTemplate(signPage, prepareForSignatureModal, templatesPage, createNewTemplatePage, successModal);
 
         await signPage.sideMenu.clickTemplates();
         await templatesPage.table.clickFirstOptionsBtn();
@@ -117,7 +120,6 @@ test.describe('Templates', () => {
 
         await step('Verify the new name of Template is visible in the table', async () => {
             await expect(await templatesPage.table.objectTitle).toHaveText(EDIT_TEMPLATE_DATA.nameField);
-
         });
 
         await step('Verify the toast message "Document successfully saved!"', async () => {
@@ -135,7 +137,7 @@ test.describe('Templates', () => {
         prepareForSignatureModal,
         templatesPage,
         createNewTemplatePage,
-        successModal
+        successModal,
     }) => {
         await description('Objective: To verify the process of duplicate template.');
         await severity(Severity.CRITICAL);
@@ -146,13 +148,10 @@ test.describe('Templates', () => {
 
         test.setTimeout(250 * 1000);
 
-        await createTemplate(signPage,
-            prepareForSignatureModal,
-            templatesPage,
-            createNewTemplatePage);
+        await createTemplate(signPage, prepareForSignatureModal, templatesPage, createNewTemplatePage, successModal);
 
         await signPage.sideMenu.clickTemplates();
-        await templatesPage.table.clickFirstOptionsBtn()
+        await templatesPage.table.clickFirstOptionsBtn();
         await templatesPage.table.clickDuplicateBtn();
         await successModal.clickOkBtn();
         await templatesPage.toast.waitForToastIsHiddenByText(TOAST_MESSAGE.templateDuplicate);
@@ -166,14 +165,14 @@ test.describe('Templates', () => {
         });
     });
 
-
     test('TC_07_29_01 | Verify that the user can delete the template', async ({
         createBusinessUserAndLogin,
         signPage,
         prepareForSignatureModal,
         templatesPage,
         createNewTemplatePage,
-        confirmDeletionModal
+        confirmDeletionModal,
+        successModal,
     }) => {
         await description('Objective: Verify that the user can successfully delete a template.');
         await severity(Severity.CRITICAL);
@@ -184,14 +183,10 @@ test.describe('Templates', () => {
 
         test.setTimeout(250 * 1000);
 
-        await createTemplate(signPage,
-            prepareForSignatureModal,
-            templatesPage,
-            createNewTemplatePage,
-            );
+        await createTemplate(signPage, prepareForSignatureModal, templatesPage, createNewTemplatePage, successModal);
 
         await signPage.sideMenu.clickTemplates();
-        await templatesPage.table.clickFirstOptionsBtn()
+        await templatesPage.table.clickFirstOptionsBtn();
         await templatesPage.table.clickDeleteBtn();
         await confirmDeletionModal.clickYesDelete();
         await templatesPage.toast.waitForToastIsHiddenByText(TOAST_MESSAGE.templateDelete);
@@ -199,7 +194,6 @@ test.describe('Templates', () => {
         await step('Verify that the number of templates in the table is 0', async () => {
             await expect(await templatesPage.table.objectTitle).toHaveCount(0);
         });
-
     });
 
 });
